New file for SwipeButtonWithImage.
